[FIX] Revision r1231 was rolled back as it broke the behavior of the Matlab/Octave function btkWriteAcquisition.
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@1300 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Code/IO/btkCLBFileIO.cpp
@@ -381,6 +381,6 @@ namespace btk
    * Constructor.
    */
   CLBFileIO::CLBFileIO()
-  : AcquisitionFileIO(AcquisitionFileIO::Binary, AcquisitionFileIO::IEEE_LittleEndian, AcquisitionFileIO::Float)
+  : AcquisitionFileIO(AcquisitionFileIO::Binary, AcquisitionFileIO::IEEE_LittleEndian, AcquisitionFileIO::Integer)
   {};
 };
